+++ Hermes_UI/history/20241004-212008_test154.txt
@@ -1,0 +1,7 @@
+{"Device": "test154", "Text": "import json", "Timestamp": "\"timestamp\":\"October 04, 2024 at 09:20PM\"", "channel": null}
+
+Processing message '{"Device": "test154", "Text": "import json", "Timestamp": "\"timestamp\":\"October 04, 2024 at 09:20PM\"", "channel": null}' against rule dictionary
+Message 'import json' matched pattern '.*'
+Rule 'Say' is inactive. Skipping.
+Message 'import json' matched pattern '.*'
+Running action 'powershell.exe -noprofile -executionpolicy bypass -file receivedMessage.ps1' in directory './actions/win10'
