+++ Hermes_UI/templates/index.html
@@ -9,7 +9,11 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/4.0.0/socket.io.js"></script>
     <script src="/static/js/formBuilder.js"></script>
     <script src="/static/js/jsonBuilder.js"></script>
-    <script src="/static/js/main.js"></script> 
+    <script src="/static/js/main.js"></script>
+    <script src="/static/js/rules.js"></script>
+    <script src="/static/js/history.js"></script>
+    <script src="/static/js/socket.js"></script>
+
 </head>
 <body>
     <div class="container mt-5">
